requirements.txt
discord.py>=2.3
aioredis>=2.0.1